requirements.txt
opencv-python
Pillow
pyzbar
requests
tqdm
